(routing) rename: renamed folders to ungroup user routes
