+++ src/app/(user)/cdp/[study]/dashboard/page.tsx
