+++ src/app/(user)/cdp/[study]/docs/dustbin.tsx
