+++ src/app/(user)/cdp/[study]/docs/file-explorer.tsx
