+++ src/app/(user)/cdp/[study]/docs/new-file.tsx
