+++ src/app/(user)/cdp/[study]/docs/page.tsx
