+++ src/app/(user)/cdp/[study]/mri/figures.tsx
