+++ src/app/(user)/cdp/[study]/mri/inner.tsx
