+++ src/app/(user)/cdp/[study]/mri/render.tsx
