+++ src/app/(user)/cdp/[study]/notes/note.tsx
