+++ src/app/(user)/cdp/[study]/notes/page.tsx
