+++ src/app/(user)/suivi/mri-validation/[[...study]]/error.tsx
