+++ src/app/(user)/suivi/mri-validation/[[...study]]/page.tsx
@@ -1,6 +1,6 @@
 import { Box, BoxContent, BoxHeader, BoxTitle } from '@/components/boxes/boxes';
-import { adminDisplay, MriFormType } from '@/app/(user)/(cdp)/[study]/mri/form/schema';
-import { RenderMRI } from '@/app/(user)/(cdp)/[study]/mri/render';
+import { adminDisplay, MriFormType } from '@/app/(user)/cdp/[study]/mri/form/schema';
+import { RenderMRI } from '@/app/(user)/cdp/[study]/mri/render';
 import { getStudyInfosWithMRI } from '../action';
 import { Mri } from '@prisma/client';
