+++ src/app/(user)/suivi/mri-validation/layout.tsx
